[backend] remove deprecated code for metrics readers (#6292)
