+++ opencti-platform/opencti-graphql/src/back.js
@@ -6,19 +6,17 @@ import { Resource } from '@opentelemetry/resources';
 import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http';
 import nconf from 'nconf';
 import { ZipkinExporter } from '@opentelemetry/exporter-zipkin';
-import { PrometheusExporter } from '@opentelemetry/exporter-prometheus';
-import { PeriodicExportingMetricReader } from '@opentelemetry/sdk-metrics';
-import { OTLPMetricExporter } from '@opentelemetry/exporter-metrics-otlp-http';
 // region static graphql modules, need to be imported before everything
 import './modules/index';
 // import managers
 import './manager/index';
+// import tracing
+import './config/tracing';
 // endregion
 import { SEMRESATTRS_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 import { platformStart } from './boot';
-import { ENABLED_EVENT_LOOP_MONITORING, ENABLED_METRICS, ENABLED_TRACING, logApp } from './config/conf';
+import { ENABLED_EVENT_LOOP_MONITORING, ENABLED_TRACING, logApp } from './config/conf';
 import { isNotEmptyField } from './database/utils';
-import { meterManager, meterProvider } from './config/tracing';
 
 // -- Apply telemetry
 // ------- Tracing
@@ -42,23 +40,6 @@ if (ENABLED_TRACING) {
   }
   // Registration
   provider.register();
-}
-// ------- Metrics
-if (ENABLED_METRICS) {
-  // OTLP - JAEGER ...
-  const exporterOtlp = nconf.get('app:telemetry:metrics:exporter_otlp');
-  if (isNotEmptyField(exporterOtlp)) {
-    const metricExporter = new OTLPMetricExporter({ url: exporterOtlp, headers: {}, concurrencyLimit: 1 });
-    meterProvider.addMetricReader(new PeriodicExportingMetricReader({ exporter: metricExporter, exportIntervalMillis: 1000 }));
-  }
-  // PROMETHEUS
-  const exporterPrometheus = nconf.get('app:telemetry:metrics:exporter_prometheus');
-  if (isNotEmptyField(exporterPrometheus)) {
-    const exporter = new PrometheusExporter({ port: exporterPrometheus });
-    meterProvider.addMetricReader(exporter);
-  }
-  // Register metrics
-  meterManager.registerMetrics();
 }
 // ------- Event loop monitoring
 if (ENABLED_EVENT_LOOP_MONITORING) {
